fixed lint issues.

+++ torch/_decomp/decompositions.py
@@ -5143,20 +5143,38 @@ def bernoulli(
     generator: Optional[torch.Generator] = None,
 ) -> torch.Tensor:
     if generator is None:
-        raw_p = torch.rand(self.size(), dtype = self.dtype, layout =  self.layout, device = self.device)
+        raw_p = torch.rand(
+            self.size(), dtype=self.dtype, layout=self.layout, device=self.device
+        )
     else:
-        raw_p = torch.rand(self.size(), generator = generator, dtype = self.dtype, layout =  self.layout, device = self.device)
+        raw_p = torch.rand(
+            self.size(),
+            generator=generator,
+            dtype=self.dtype,
+            layout=self.layout,
+            device=self.device,
+        )
     p = (raw_p < self).to(self.dtype)
     return p
+
 
 @register_decomposition(aten.bernoulli.p)
 def bernoulli_p(self, p, *, generator: Optional[torch.Generator] = None):
     if generator is None:
-        raw_p = torch.rand(self.size(), dtype = self.dtype, layout =  self.layout, device = self.device)
+        raw_p = torch.rand(
+            self.size(), dtype=self.dtype, layout=self.layout, device=self.device
+        )
     else:
-        raw_p = torch.rand(self.size(), generator = generator, dtype = self.dtype, layout =  self.layout, device = self.device)
+        raw_p = torch.rand(
+            self.size(),
+            generator=generator,
+            dtype=self.dtype,
+            layout=self.layout,
+            device=self.device,
+        )
     p = (raw_p < p).to(self.dtype)
     return p
+
 
 def isin_default(elements, test_elements, *, invert=False):
     if elements.numel() == 0:
